fix(Desktop): ウィンドウ位置が保存されないのを修正
