+++ lib/providers.dart
@@ -33,6 +33,7 @@ import "package:miria/repository/time_line_repository.dart";
 import "package:miria/repository/user_list_time_line_repository.dart";
 import "package:miria/router/app_router.dart";
 import "package:miria/util/file_system_io.dart" as fs;
+import "package:miria/util/window_listener.dart";
 import "package:misskey_dart/misskey_dart.dart";
 import "package:riverpod/riverpod.dart";
 import "package:riverpod_annotation/riverpod_annotation.dart";
@@ -177,6 +178,9 @@ CacheManager cacheManager(Ref ref) => CacheManager(
     fileSystem: fs.IOFileSystem("libCachedImageData"),
   ),
 );
+
+@Riverpod(keepAlive: true)
+MiriaWindowListener miriaWindowListener(Ref ref) => MiriaWindowListener(ref);
 
 @freezed
 abstract class AccountContext with _$AccountContext {
